Conflicts:
	mcwrap.m

+++ examples/example1_cpp/reverse_it.cpp
@@ -1,6 +1,6 @@
 #include "reverse_it.h"
 
-void reverse_it(int N,float *X_out,float *X_in) {
+void reverse_it(int N,double *X_out,double *X_in) {
     for (int i=0; i<N; i++) {
         X_out[N-1-i]=X_in[i];
     }
